1999-06-24  Norbert Warmuth  <nwarmuth@privat.circular.de>
* gnome/gmc-chargrid.c (update_strip): Paint every character with the
assigned fore and background color instead of using the colors of
the first character for the whole line.


1999-06-24  smil@linuxfan.com

* src/view.c (view_update_bytes_per_line, display): show correctly the
offset of the file (full 8 digit), display offset in bold colour

(view_labels): use goto_addr instead of goto_line in hex mode

(goto_addr): New function. Goto offset address in hex mode.

* src/view.h: use unsigned long for hexedit cursor position in file

+++ src/view.h
@@ -39,7 +39,7 @@ typedef struct {
 				/* For the case of WINCH we should reset it to -1 */
     unsigned long start_display;/* First char displayed */
     int  start_col;		/* First displayed column, negative */
-    unsigned int  edit_cursor;  /* HexEdit cursor position in file */
+    unsigned long edit_cursor;  /* HexEdit cursor position in file */
     char hexedit_mode;          /* Hexidecimal editing mode flag */ 
     char nib_shift;             /* A flag for inserting nibbles into bytes */
     enum ViewSide view_side;	/* A flag for the active editing panel */
